Moved AuthProvider to src/providers/
